+++ kuStudy/kuStudy/view controller/LibraryViewController.swift
@@ -38,6 +38,7 @@ class LibraryViewController: UIViewController {
         tableView.dataSource = self
         tableView.emptyDataSetDelegate = self
         tableView.emptyDataSetSource = self
+        tableView.contentInset = UIEdgeInsets(top: 0, left: 0, bottom: 49, right: 0)
         tableView.tableFooterView = UIView()
         tableView.addSubview(refreshControl)
         refreshControl.addTarget(self, action: #selector(updateData(_:)), forControlEvents: .ValueChanged)