+++ kuStudy/kuStudy/view controller/SummaryViewController.swift
@@ -34,6 +34,7 @@ class SummaryViewController: UIViewController, UIViewControllerPreviewingDelegat
         tableView.dataSource = self
         tableView.emptyDataSetDelegate = self
         tableView.emptyDataSetSource = self
+        tableView.contentInset = UIEdgeInsets(top: 0, left: 0, bottom: 49, right: 0)
         tableView.tableFooterView = UIView()
         tableView.addSubview(refreshControl)
         refreshControl.addTarget(self, action: #selector(updateData(_:)), forControlEvents: .ValueChanged)